Use any registered TableNameResolver in getOverriddenTableName.
This adds consideration of any TableNameResolver in the dynamoDBMapperConfig in
the getOverriddenTableName method. Without this change it is not possible to run
certain types of queries against tables whose names are not known at build time.
Registering a TableNameResolver did not help since it is a mapper concept and
these queries are run using QueryRequests directly through the client.